First pass to remove dependency on Python in arcticdb_core
