+++ cpp/arcticdb/column_store/memory_segment.hpp
@@ -126,15 +126,8 @@ public:
         impl_->init_column_map();
     }
 
-    template<class T, template<class> class Tensor>
-    requires std::integral<T> || std::floating_point<T>
-    void set_array(position_t pos, Tensor<T> &val) {
-        impl_->set_array(pos, val);
-    }
-
-    template<class T>
-    requires std::integral<T> || std::floating_point<T>
-    void set_array(position_t pos, py::array_t<T>& val) {
+    template<arithmetic_tensor TensorType>
+    void set_array(position_t pos, TensorType &val) {
         impl_->set_array(pos, val);
     }
