+++ cpp/arcticdb/column_store/memory_segment_impl.hpp
@@ -9,6 +9,7 @@
 
 #include <arcticdb/entity/types.hpp>
 #include <arcticdb/column_store/column.hpp>
+#include <arcticdb/util/concepts.hpp>
 #include <arcticdb/util/offset_string.hpp>
 #include <arcticdb/util/preconditions.hpp>
 
@@ -506,17 +507,8 @@ public:
         set_string(idx, val);
     }
 
-    template<class T, template<class> class Tensor>
-    requires std::integral<T> || std::floating_point<T>
-    void set_array(position_t pos, Tensor<T> &val) {
-        magic_.check();
-        ARCTICDB_SAMPLE(MemorySegmentSetArray, 0)
-        column_unchecked(pos).set_array(row_id_ + 1, val);
-    }
-
-    template<class T>
-    requires std::integral<T> || std::floating_point<T>
-    void set_array(position_t pos, py::array_t<T>& val) {
+    template<arithmetic_tensor TensorType>
+    void set_array(position_t pos, TensorType &val) {
         magic_.check();
         ARCTICDB_SAMPLE(MemorySegmentSetArray, 0)
         column_unchecked(pos).set_array(row_id_ + 1, val);
