+++ cpp/arcticdb/storage/library_manager.hpp
@@ -13,6 +13,10 @@
 #include <arcticdb/storage/storage_override.hpp>
 #include <arcticdb/util/lru_cache.hpp>
 
+// TODO: clean and remove this
+#include <pybind11/pybind11.h>
+namespace py = pybind11;
+
 namespace arcticdb::storage {
     enum class ModifiableLibraryOption {
         DEDUP = 1,
