+++ cpp/arcticdb/util/flatten_utils.hpp
@@ -8,27 +8,15 @@
 #pragma once
 
 #include <arcticdb/entity/types.hpp>
+#include <arcticdb/util/concepts.hpp>
 #include <arcticdb/util/preconditions.hpp>
-
-#include <pybind11/numpy.h>
-
-namespace py = pybind11;
 
 namespace arcticdb::util {
 
 using namespace arcticdb::entity;
 
-template<class T, template<class> class Tensor>
-inline bool has_funky_strides(Tensor<T> &a) {
-    for (ssize_t i = 0; i < a.ndim(); ++i) {
-        if (a.strides(i) < 0 || a.strides(i) % a.itemsize() != 0)
-            return true;
-    }
-    return false;
-}
-
-template<class T>
-inline bool has_funky_strides(py::array_t<T>& a) {
+template<tensor TensorType>
+inline bool has_funky_strides(TensorType& a) {
     for (ssize_t i = 0; i < a.ndim(); ++i) {
         if (a.strides(i) < 0 || a.strides(i) % a.itemsize() != 0)
             return true;
@@ -57,33 +45,25 @@ struct stride_advance_optimistic {
     }
 };
 
-template<class T, template<class> class Tensor>
-auto shape_and_strides(Tensor<T> &array, ssize_t dim) {
+template<tensor TensorType>
+auto shape_and_strides(TensorType &array, ssize_t dim) {
     auto total_dim = array.ndim();
     shape_t sh = array.shape(total_dim - size_t(dim));
     stride_t sd = array.strides(total_dim - size_t(dim));
     return std::make_pair(sh, sd);
 }
 
-template<class T>
-auto shape_and_strides(py::array_t<T>& array, ssize_t dim) {
-    auto total_dim = array.ndim();
-    shape_t sh = array.shape(total_dim - size_t(dim));
-    stride_t sd = array.strides(total_dim - size_t(dim));
-    return std::make_pair(sh, sd);
-}
-
-template<class T, template<class> class Tensor, typename AdvanceFunc>
+template<tensor TensorType, typename AdvanceFunc>
 class FlattenHelperImpl {
-    Tensor<T> &array_;
+    TensorType &array_;
     AdvanceFunc advance_func_;
 
 public:
-    explicit FlattenHelperImpl(Tensor<T> &a) : array_(a) {}
+    explicit FlattenHelperImpl(TensorType &a) : array_(a) {}
 
-    using raw_type = T;
+    using raw_type = typename TensorType::value_type;
 
-    void flatten(T *&dest, const T *src, ssize_t dim) const {
+    void flatten(raw_type*& dest, const raw_type* src, ssize_t dim) const {
         auto [sh, sd] = shape_and_strides(array_, dim);
 
         for (shape_t i = 0; i < sh; ++i) {
@@ -97,21 +77,21 @@ public:
     }
 };
 
-template<class T, template<class> class Tensor>
+template<tensor TensorType>
 class FlattenHelper {
-    Tensor<T> &array_;
+    TensorType &array_;
 
   public:
-    explicit FlattenHelper(Tensor<T> &a) : array_(a) {}
+    explicit FlattenHelper(TensorType &a) : array_(a) {}
 
-    using raw_type = T;
+    using raw_type = typename TensorType::value_type;
 
-    void flatten(T *&dest, const T *src) const {
+    void flatten(raw_type*& dest, const raw_type* src) const {
         if(has_funky_strides(array_)) {
-            FlattenHelperImpl<T, Tensor, stride_advance_conservative<T>> flh{array_};
+            FlattenHelperImpl<TensorType, stride_advance_conservative<raw_type>> flh{array_};
             flh.flatten(dest, src, array_.ndim());
         } else {
-            FlattenHelperImpl<T, Tensor, stride_advance_optimistic<T>> flh{array_};
+            FlattenHelperImpl<TensorType, stride_advance_optimistic<raw_type>> flh{array_};
             flh.flatten(dest, src, array_.ndim());
         }
     }
